[life-xp-activity-service -26] implement getActivities() method with E2E Test
- refactoring

+++ life-xp-activity-service-core/src/main/java/com/bladestepapp/lifexpactivityservicecore/domain/Activity.java
@@ -17,14 +17,14 @@ public class Activity {
 
     private ActivityUnit unit;
 
-    private double baseXp;
+    private int baseXp;
 
     public static Activity create(
             String name,
             String description,
             ActivityCategory category,
             ActivityUnit unit,
-            double baseXp
+            int baseXp
     ) {
         return new Activity(name,
                 description,
